Added /LZWDecode filter.
This is a crazy thing. My test files (from years ago) are on another
computer (I do not want to get confused by my old sources).  It's about
impossible to find a fitting PDF for this task, but I know it is important
for older PDF files.  So this change is very experimental.  I've checked the
things to work with some hacky methods and an error report for a Java
library that had such a wanted PDF included.