Started sort of work on fonts.
The fontnamemap.txt contains file names on my system and the fitting
Postscript names of the font. Most fonts have to be included into the SVG as
SVG font. To avoid calling fontforge from the program, I want for now to use
my systems equivalents of some known fonts. This has to be replaced by a
better solution of course.